drsmith-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-summernote>=0.8.20.0
django-contrib-comments>=2.2.0
django-comments-xtd>=2.9.9
Pillow>=10.0.0
requests>=2.31.0